Models/requirements.txt
#LangChain Core
langchain
langchain-Core

#OpenAI Integration
langchain-OpenAI
Openai

#Anthropic Integration
langchain-anthropic

#Google Gemini Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Environment Variable Management
python-dotenv

#ML Utilities
numpy
scikit-learn

